Move persistence and inbox messaging contracts out of EF Core
The persistence contracts (IUnitOfWork, the inbox dedup and consumer
checkpoint store SPIs) and the inbox messaging contracts (the dispatch
seam, envelope, and outcome) were declared in the Trellis.EntityFrameworkCore
namespace and shipped from the EF packages, so implementing any of them —
or reusing the standard transactional commit pipeline — forced a dependency
on EF Core even though none of the contracts names an EF type.

Implement the relocation (ADR-006):

- New Trellis.Core-only package Trellis.Persistence.Abstractions hosts
  IUnitOfWork, IInboxStore, IConsumerCheckpointStore, and a new persistence-
  native InboxRecord. A non-EF adapter can implement these without referencing
  EF Core.
- IInboxStore.TryRecordAsync takes InboxRecord instead of the messaging
  IntegrationEnvelope, so the store contract no longer depends on a messaging
  type. The EF dispatcher maps the envelope to a record; the EF row persists
  the same columns as before.
- The inbox dispatch contracts (IInboxDispatcher, IntegrationEnvelope,
  InboxDispatchOutcome) and the (already store-agnostic) TransactionalCommandBehavior
  move into Trellis.Mediator, alongside the sibling integration-event contracts.
  A new provider-neutral AddTransactionalCommandBehavior() installs the commit
  pipeline; the EF adapter's AddTrellisUnitOfWork calls it.
- Mediator now detects the transactional behavior by type rather than by a
  hard-coded EF type-name string.

All moved types keep the flat Trellis namespace (store contracts) or the
Trellis.Mediator namespace (dispatch contracts); EF Core remains one adapter.
