+++ Trellis.EntityFrameworkCore.Inbox/src/Trellis.EntityFrameworkCore.Inbox.csproj
@@ -17,6 +17,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Trellis.EntityFrameworkCore\src\Trellis.EntityFrameworkCore.csproj" />
     <ProjectReference Include="..\..\Trellis.Mediator\src\Trellis.Mediator.csproj" />
+    <ProjectReference Include="..\..\Trellis.Persistence.Abstractions\src\Trellis.Persistence.Abstractions.csproj" />
   </ItemGroup>
   <ItemGroup>
     <InternalsVisibleTo Include="Trellis.EntityFrameworkCore.Inbox.Tests" />
